Remove need for UrlUtil and downgrade Kotlin ver to 1.4.21 to ensure IDE compatibility

+++ appvox-core/src/main/kotlin/dev/fabiou/appvox/core/configuration/Constant.kt
@@ -7,5 +7,5 @@ internal object Constant {
      * Minimum request delay is set in an effort to provide a polite scraper that do not affect the scraped service
      * An @exception will be thrown
      */
-    const val MIN_REQUEST_DELAY: Long = 500
+    const val MIN_REQUEST_DELAY: Int = 500
 }
